Corrigindo adição de arquivos desnecessários
